requirements.txt
fastapi
groq
python-dotenv
requests
streamlit
uvicorn
